+++ tailwind.config.js
@@ -1,22 +1,5 @@
-theme1 = {
-  dark: "#132743",
-  light: "whitesmoke",
-  primary: "#edc988",
-  secondary: "#f8efd4",
-  accent: "#d7385e",
-  muted: "#ddd",
-};
-
-const breakpoints = {
-  sm: 500,
-  md: 800,
-  lg: 1080,
-  xl: 1366,
-};
-
-function px(num) {
-  return num + "px";
-}
+const { defaultBreakpoints, generateScreens } = require("./breakpoints");
+const { defaultTheme } = require("./themes");
 
 module.exports = {
   future: {
@@ -25,12 +8,7 @@ module.exports = {
   },
   purge: ["./src/components/**/*.tsx", "./src/pages/**/*.tsx"],
   theme: {
-    screens: {
-      sm: { max: px(breakpoints.sm) },
-      md: { min: px(breakpoints.sm + 1), max: px(breakpoints.md) },
-      lg: { min: px(breakpoints.md + 1), max: px(breakpoints.lg) },
-      xl: { min: px(breakpoints.lg + 1), max: px(breakpoints.xl) },
-    },
+    screens: generateScreens(defaultBreakpoints),
     fontFamily: {
       serif: ["David Libre", "Georgia", "Times New Roman", "Times", "serif"],
       cursive: ["Sansita Swashed", "cursive"],
@@ -43,14 +21,7 @@ module.exports = {
       height: {
         "screen/2": "50vh",
       },
-      colors: {
-        dark: "#132743",
-        light: "whitesmoke",
-        primary: "#edc988",
-        secondary: "#f8efd4",
-        accent: "#d7385e",
-        muted: "#aaa",
-      },
+      colors: defaultTheme,
     },
   },
   variants: {},
