+++ themes.js
@@ -1,0 +1,15 @@
+module.exports = {
+  darkTheme: {
+    accent: "#232931",
+    dark: "#41aea9",
+    light: "#eeeeee",
+  },
+  defaultTheme: {
+    dark: "#132743",
+    light: "whitesmoke",
+    primary: "#edc988",
+    secondary: "#f8efd4",
+    accent: "#d7385e",
+    muted: "#aaa",
+  },
+};
